refactor attributes, fix stuff i moved incorrectly
