+++ ffmpeg-gui/src/main/java/attributes/AudioAttributes.java
@@ -1,7 +1,5 @@
-package entity;
+package attributes;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import net.bramp.ffmpeg.probe.FFmpegStream;
