+++ ffmpeg-gui/src/main/java/entity/VideoJob.java
@@ -1,5 +1,7 @@
 package entity;
 
+import attributes.AudioAttributes;
+import attributes.VideoAttributes;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
