+++ ffmpeg-gui/src/main/java/interface_adapter/convert_video_file/ConvertVideoFileViewModel.java
@@ -1,6 +1,6 @@
 package interface_adapter.convert_video_file;
 
-import entity.AudioCodec;
+import constant.AudioCodec;
 import constant.VideoCodec;
 import constant.VideoFormat;
 import interface_adapter.ViewModel;
